+++ server/api/auth/config.js
@@ -1,44 +1,7 @@
-const expressSession = require('express-session');
-const passport = require('passport');
-const bcrypt = require('bcrypt-nodejs');
-const flash = require('connect-flash');
-const LocalStrategy = require('passport-local').Strategy;
-
-const dbQry = require('./db/dbQueries');
-
-module.exports = (app) => {
-  app.use(expressSession({ secret: 'keyboard cat' }));
-  app.use(passport.initialize());
-  app.use(passport.session());
-  app.use(flash());
-
-  passport.use(new LocalStrategy(
-    {
-      usernameField: 'email',
-      passwordField: 'password',
-    },
-    (email, password, done) => {
-      dbQry.signIn(email, password, (err, result) => {
-        if (err) { return done(err); } // error code 500
-        if (!result) {
-          console.log('User does not exist');
-          return done(null, false, { message: 'Incorrect email.' });
-        }
-        const user = result.rows[0];
-        if (!bcrypt.compareSync(password, user.password)) {
-          console.log('Incorrect password');
-          return done(null, false, { message: 'Incorrect password.' });
-        }
-        return done(null, user);
-      });
-    }
-  ));
-
-  passport.serializeUser((user, done) => {
-    done(null, user);
-  });
-
-  passport.deserializeUser((user, done) => {
-    done(null, user);
-  });
+module.exports = {
+  db: {
+    host: 'localhost',
+    port: 5432,
+    dbName: 'openab',
+  },
 };
