--- server/api/auth/db/dbConfig.js
@@ -1,7 +1,0 @@
-module.exports = {
-  db: {
-    host: 'localhost',
-    port: 5432,
-    dbName: 'openab',
-  },
-};
