--- test/server/api/controllerSpec.js
@@ -1,2 +1,0 @@
-const expect = require('chai').expect;
-const authController = require('../../../server/api/auth/controller.js');
